% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTensGS_functions.R
\name{marker.QC}
\alias{marker.QC}
\title{SNP marker cleaning and QC}
\usage{
marker.QC(
  marker.mat,
  impute = T,
  cutoff = 0.8,
  min.maf = 0.05,
  max.NAs = 0.1,
  n.cores = NULL,
  verbose = T
)
}
\arguments{
\item{marker.mat}{NxM marker matrix with row names and column names. Markers coded as 0,1,2 for hom, het,hom.}

\item{impute}{Logical whether to run imputation step. Default is TRUE}

\item{cutoff}{correlation threshold for pruning. Numeric between 0 and 1. Default is 0.8}

\item{min.maf}{Minimum minor allele frequency to keep. Numeric between 0 and 0.5. Default is 0.05}

\item{max.NAs}{Maximum proportion of NA missing data Numeric between 0 and 1. Default is 0.1}

\item{n.cores}{Number of cores that can be used to run imputation in parallel. The Default detectcores() from Parallel will use available multiple cores on a Windows OS but n.cores should be defined when running on a HPC or other OS.}

\item{verbose}{Logical whether to print progress. Default is TRUE.}
}
\value{
A NxM marker matrix.
}
\description{
Uses a SNP marker matrix and sequentially filters for missing data, imputes remaining missing NA SNP alleles per marker, filters for minor allele frequency and 'prunes' markers based on marker correlation.
}
\examples{
library(MTensGS)
data(TGdata)

QCd.markers<-marker.QC(marker.mat = TGgenos)

#Check size of new marker dataset
dim(TGgenos)
dim(QCd.markers)


}
