% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTensGS_functions.R
\name{MT.PCSI.plot}
\alias{MT.PCSI.plot}
\title{Plot multi-trait Principal component selection indices}
\usage{
MT.PCSI.plot(
  X,
  SI.traits = NULL,
  weights = NULL,
  model = "MeanMtens",
  col1 = NULL,
  col2 = NULL,
  cex1 = 0.7,
  cex2 = 1
)
}
\arguments{
\item{X}{Object from Train.Predict.MTens()}

\item{SI.traits}{\emph{vector} of trait names included in X to visualise. Default is all traits.}

\item{weights}{\emph{vector} of weightings for each trait in SI.traits. Default is equal weightings (all = 1).}

\item{model}{Prediction model type to use. Options include. Default is the MT.ensemble}

\item{col1, col2}{Colours to use for trait and genotype ID names in biplot.}
}
\description{
Plot a PCA biplot visualisation of multiple predicted traits to aid selection. Positive and negative trade offs among traits can be considered and weightings applied to subsets of desired traits.
}
\examples{
library(MTensGS)
data(TGdata)

QCd.markers<-marker.QC(marker.mat = TGgenos)

#Split the dataset in two
old.genos<-QCd.markers[1:300,]
new.genos<-QCd.markers[301:376,]
old.phens<-TGphenos[1:300,]

new_preds<-Train.Predict.MTens(Genos.train = old.genos,
                               New.Genos = new.genos,
                               Phenos = old.phens,
                               n.valid.rounds = 1,
                               n.valid.folds = 6,
                               base.models = c("rrblup","LASSO"))

sel.ind<-MT.PCSI.plot(new_preds,SI.traits = c("YLD_GBR_2011","YLD_GBR_2010","YLD_DEU_2010","PROT")
                      ,weights = c(1,1,1,2),cex1 = 0.8)

}
