% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTensGS_functions.R
\name{plot.MT.PC.SI}
\alias{plot.MT.PC.SI}
\title{Plot multi-trait Principle component selection indices}
\usage{
\method{plot}{MT.PC.SI}(
  X,
  SI.traits = NULL,
  weights = NULL,
  model = "MT Ensemble preds",
  sub.model = "RF MT Ens",
  col1 = rgb(0.1, 0.2, 0.8, 0.7, maxColorValue = 1),
  col2 = rgb(1, 0, 0, 0.6, maxColorValue = 1)
)
}
\description{
Plot multi-trait Principle component selection indices
}
