% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTensGS_functions.R
\name{Train.Predict.MTens}
\alias{Train.Predict.MTens}
\title{Train and predict multi-trait ensembles}
\usage{
Train.Predict.MTens(
  Genos.train,
  New.Genos,
  Phenos,
  n.valid.folds = 8,
  rf.ntrees = 300,
  gbm.ntrees = 200,
  n.cores = detectCores() - 1,
  base.models = c("rrblup", "EGBLUP", "RKHS", "LASSO", "RF", "GBM", "XGB")
)
}
\arguments{
\item{Genos.train}{NxM marker \emph{matrix} for genotypes in the training set with row names and column names as output from marker.QC(). Markers coded as 0,2 for each SNP allele. Inbred lines are assumed. Must have no missing (NA)
values. Row names are Genotype ID names and column are SNP marker names.}

\item{New.Genos}{NxM marker \emph{matrix} for new genotypes to be predicted with row names and column names as output from marker.QC(). Markers coded as 0,2 for each SNP allele. Inbred lines are assumed. Must have no missing (NA)
values. Row names are Genotype ID names and column are SNP marker names.}

\item{Phenos}{Data frame of N x p trait data for genotypes in the training set. Row names are Genotype ID matching row names of Genos. Column names are trait names. Missing data is allowed.}

\item{n.valid.folds}{(integer) number of validation folds to run within each test fold. Default if 8.}

\item{rf.ntrees}{(integer) Number of trees to use for each random forest model}

\item{gbm.ntrees}{(integer) Max number of trees to use for each GBM model}

\item{n.cores}{Number of cores that can be used to run cross validation of RF, GBM and XGB models. The Default detectcores() from Parallel will use available multiple
cores on a Windows OS but n.cores should be defined when running on a HPC or other OS.}

\item{base.models}{\emph{vector} of base model names to use. Options include:
c("rrblup","EGBLUP","RKHS","LASSO","RF","GBM","XGB")

-\strong{rrblup} Ridge regression/GBLUP implemented in the rrblup package using a linear GRM.

-\strong{EGBLUP} The Extended GBLUP incorporating a linear and epistatic GRM kernel implemented in the BGLR package.

-\strong{RKHS} Replicating Kernel Hilbert Space fitting multiple Gaussian kernels using the kernel averaging method for several bandwidth parameters. Implemented in the BGLR package.

-\strong{LASSO} Linear Shrinkage model using the LASSO penalty to shrink most SNP effects to 0. Implemented in the glmnet package. Each model is cross validated with several values of lambda.

-\strong{RF} Random forest decision tree ensemble learning method implemented in the randomForest package. Forests of 300 trees are grown as defualt but can be increased using rf.ntrees if needed.

-\strong{GBM} Gradient Boosting Machine learning models implemented in the gbm package. May take a long time... 5 fold cross validation is performed on each model to
determine optimum tree iteration to use. cross validations can be run in parallel on multiple cores by setting n.cores if not defined by default.

-\strong{XGB} Extreme Gradient Boosting machine learning implemented in the xgboost package. A hyperparameter grid search is performed accross different tree depth and shrinkage
parameters and each model is 5-fold cross validated  to determine optimum tree iteration. May take a long time...}

\item{CV.groups}{(list) of vectors containing Genotype ID names. Cross validation can also be performed across predefined subsets of genotypes.}

\item{verbose}{Level of progress to print throughout. 0 = none; 1 = some; 2 = all.}
}
\value{
Returns a list object of:

-\strong{All out-of-fold predictions} for all genotypes for base prediction models, genetic model ensemble and multi-trait ensemble models for each trait.

-\strong{Phenotype data} all observed multi-trait input data.

-\strong{Prediction accuracies} list of correlation coeficient (r) for each model for each trait.

-\strong{Base models} \emph{vector} of base prediction models used.

-\strong{Traits} \emph{vector} of trait names used.

-\strong{CV fold allocations} A two level list of length = n.fold.rounds, each containing the genotype ID names allocated to each test fold.

#' @details Ensemble predictions for multiple genetic models and multiple traits are trained on out-of-validation fold predictions using several model approaches.
Including a range of genetic models that capture a diversity of genetic architectures may improve accuracy of ensembles. Unbalanced phenotyping across
genotypes may increase accuracy of multi-trait ensembles when several related but partially phenotyped traits are included.
}
\description{
Trains several base genomic prediction models as well as multiple genetic model and multi-trait ensemble prediction models. Minimum input requirements are SNP genotype data and multi-trait phenotype data base prediction models should run well on default parameters.
}
