% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTensGS_functions.R
\name{Train.Predict.MTens}
\alias{Train.Predict.MTens}
\title{Train and predict multi-trait ensembles}
\usage{
Train.Predict.MTens(
  Genos.train,
  New.Genos,
  Phenos,
  n.valid.folds = 8,
  n.valid.rounds = 3,
  rf.ntrees = 400,
  gbm.ntrees = 300,
  n.cores = NULL,
  bglr.saveAt = NULL,
  base.models = c("rrblup", "BayesB", "BayesC", "EGBLUP", "RKHS", "LASSO", "RF", "GBM"),
  verbose = 2
)
}
\arguments{
\item{Genos.train}{NxM marker \emph{matrix} for genotypes in the training set with row names and column names as output from marker.QC(). Markers coded as 0,2 for each SNP allele. Inbred lines are assumed. Must have no missing (NA)
values. Row names are Genotype ID names and column are SNP marker names.}

\item{New.Genos}{NxM marker \emph{matrix} for new genotypes to be predicted with row names and column names as output from marker.QC(). Markers coded as 0,2 for each SNP allele. Inbred lines are assumed. Must have no missing (NA)
values. Row names are Genotype ID names and column are SNP marker names.}

\item{Phenos}{Data frame of N x p trait data for genotypes in the training set. Row names are Genotype ID matching row names of Genos. Column names are trait names. Missing data is allowed.}

\item{n.valid.folds}{(integer) number of validation folds to run for each trait and base model. Default is 8.}

\item{rf.ntrees}{(integer) Number of trees to use for each Random forest model}

\item{gbm.ntrees}{(integer) Max number of trees to use for each GBM model}

\item{n.cores}{Number of cores that can be used to run cross validation of RF, GBM and XGB models. The Default detectcores() from Parallel will use available multiple
cores on a Windows OS but n.cores should be defined when running on a HPC or other OS.}

\item{bglr.saveAt}{file path location to save temporary data files for 'BGLR' model functions. Using a local folder may speed up computing time. Default is the current working directory.}

\item{base.models}{\emph{vector} of base model names to use. Options include:
c("rrblup","EGBLUP","RKHS","LASSO","RF","GBM")

-\strong{rrblup} Ridge regression/GBLUP implemented in the rrblup package using a linear GRM.

-\strong{EGBLUP} The Extended GBLUP incorporating a linear and epistatic GRM kernel implemented in the BGLR package.

-\strong{BayesB} Bayesian linear regression of marker effects with point of mass at zero plus t-slab priors implemented in the BGLR package.

-\strong{BayesC} Bayesian linear regression of marker effects with point of mass at zero plus Gaussian slab priors implemented in the BGLR package.

-\strong{RKHS} Replicating Kernel Hilbert Space fitting multiple Gaussian kernels using the kernel averaging method for several bandwidth parameters. Implemented in the BGLR package.

-\strong{LASSO} Linear Shrinkage model using the LASSO penalty to shrink most SNP effects to 0. Implemented in the glmnet package. Each model is cross validated with several values of lambda.

-\strong{RF} Random forest decision tree ensemble learning method implemented in the randomForest package. Forests of 300 trees are grown as defualt but can be increased using rf.ntrees if needed.

-\strong{GBM} Gradient Boosting Machine learning models implemented in the gbm package. May take a long time... 5 fold cross validation is performed on each model to
determine optimum tree iteration to use. cross validations can be run in parallel on multiple cores by setting n.cores if not defined by default.}

\item{verbose}{Level of progress to print throughout. 0 = none; 1 = some; 2 = all.}

\item{n.CV.rounds}{\emph{integer} number of rounds of k-fold cross validation for test folds to perform. Default is 2}
}
\value{
Returns a list object of:

-\strong{All new predictions} \emph{array} Predictions for all new genotypes for base prediction models. Array with dimensions of genotypes,  base models and traits.

-\strong{Traits} \emph{vector} of trait names used.

-\strong{Base models} \emph{vector} of base prediction models used.
}
\description{
Trains several base genomic prediction models as well as multiple genetic model and multi-trait ensemble prediction models. Minimum input requirements are bi-alleleic  marker genotype data and multi-trait phenotype data. Base prediction models should run well on default parameters.
}
\details{
Ensemble predictions for for genotypes in New.Genos for multiple genetic models and multiple traits are made after trained on out-of-validation fold predictions using several model
approaches in the Genos.train and Phenos objects. Including a range of genetic models that capture a diversity of genetic architectures may improve accuracy of ensembles but running all available base models takes a long time. Unbalanced phenotyping across
genotypes may increase accuracy of multi-trait ensembles when several related but partially phenotyped traits are included.
}
\examples{
library(MTensGS)
data(TGdata)

QCd.markers<-marker.QC(marker.mat = TGgenos)

#Split the dataset in two
old.genos<-QCd.markers[1:300,]
new.genos<-QCd.markers[301:376,]
old.phens<-TGphenos[1:300,]

new_preds<-Train.Predict.MTens(Genos.train = old.genos,
                               New.Genos = new.genos,
                               Phenos = old.phens,
                               n.valid.rounds = 1,
                               n.valid.folds = 6,
                               base.models = c("rrblup","LASSO"))



}
